Moved animation to the AnimatedTabGroup class
