+++ jvm/workbookapp/src/main/kotlin/org/wycliffeassociates/otter/jvm/workbookapp/ui/chromeablestage/tabgroups/ChapterTabGroup.kt
@@ -3,11 +3,10 @@ package org.wycliffeassociates.otter.jvm.workbookapp.ui.chromeablestage.tabgroup
 import javafx.scene.control.Tab
 import org.wycliffeassociates.otter.common.data.model.ResourceMetadata
 import org.wycliffeassociates.otter.jvm.workbookapp.ui.cardgrid.view.CardGridFragment
-import org.wycliffeassociates.otter.jvm.workbookapp.ui.chromeablestage.AnimatedTab
 import org.wycliffeassociates.otter.jvm.workbookapp.ui.workbook.viewmodel.WorkbookViewModel
 import tornadofx.*
 
-class ChapterTabGroup : TabGroup() {
+class ChapterTabGroup : AnimatedTabGroup() {
     private val workbookViewModel: WorkbookViewModel by inject()
     private val tabMap: MutableMap<String, Tab> = mutableMapOf()
