+++ jvm/workbookapp/src/main/kotlin/org/wycliffeassociates/otter/jvm/workbookapp/ui/chromeablestage/tabgroups/RecordResourceTabGroup.kt
@@ -4,16 +4,15 @@ import org.wycliffeassociates.otter.common.data.model.ContentType
 import org.wycliffeassociates.otter.jvm.workbookapp.ui.resourcetakes.view.RecordableTab
 import org.wycliffeassociates.otter.jvm.workbookapp.ui.resourcetakes.viewmodel.RecordResourceViewModel
 import org.wycliffeassociates.otter.jvm.utils.getNotNull
-import org.wycliffeassociates.otter.jvm.workbookapp.ui.chromeablestage.AnimatedChromeableTabPane
 import tornadofx.*
 
-class RecordResourceTabGroup : TabGroup() {
+class RecordResourceTabGroup : AnimatedTabGroup() {
     private val viewModel: RecordResourceViewModel by inject()
 
     init {
         viewModel.transitionDirectionProperty.onChange {
             it?.let {
-                (tabPane as? AnimatedChromeableTabPane)?.animate(it)
+                animate(it)
                 viewModel.transitionDirectionProperty.set(null)
             }
         }
