+++ jvm/workbookapp/src/main/kotlin/org/wycliffeassociates/otter/jvm/workbookapp/ui/chromeablestage/tabgroups/TransitionDirection.kt
@@ -1,0 +1,6 @@
+package org.wycliffeassociates.otter.jvm.workbookapp.ui.chromeablestage.tabgroups
+
+enum class TransitionDirection {
+    LEFT,
+    RIGHT
+}
