+++ jvm/workbookapp/src/main/kotlin/org/wycliffeassociates/otter/jvm/workbookapp/ui/resourcetakes/view/RecordableTab.kt
@@ -4,7 +4,7 @@ import javafx.scene.paint.Color
 import org.wycliffeassociates.otter.common.domain.content.Recordable
 import org.wycliffeassociates.otter.jvm.controls.statusindicator.StatusIndicator
 import org.wycliffeassociates.otter.jvm.controls.statusindicator.statusindicator
-import org.wycliffeassociates.otter.jvm.workbookapp.ui.chromeablestage.AnimatedTab
+import org.wycliffeassociates.otter.jvm.workbookapp.ui.chromeablestage.tabgroups.AnimatedTab
 import org.wycliffeassociates.otter.jvm.workbookapp.ui.resourcetakes.viewmodel.RecordableTabViewModel
 import tornadofx.*
